+++ src/aux-server/aux-web/aux-player/PlayerGameView/PlayerGameView.vue
@@ -30,6 +30,7 @@
 
             <div class="slider-visible" @touchmove="moveTouch($event)"></div>
             <div class="side-visible"></div>
+            <div class="sideRight-visible"></div>
 
             <div
                 class="slider-hidden"